requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
starlette==0.45.3
pydantic==2.10.6
pika==1.3.2
python-dotenv==1.0.1
